chore: fix linting issues
